Completing payment section(seller side)

+++ database/migrations/2022_07_12_070109_create_payments_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Cart::class);
             $table->enum('status', ['pending', 'preparing', 'sending', 'delivered']);
-            $table->integer('totalPrice');
+            $table->float('totalPrice');
             $table->timestamps();
         });
     }
